organize sub comps in block explorer
